ScholarPathAI/ScholarPathAI/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.19.0",
    "numpy>=2.2.4",
    "openai>=1.66.3",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "streamlit>=1.43.2",
    "watchdog>=6.0.0",
]
